requirements.txt
django
django-cors-headers
django-import-export
psycopg2
django-rest-framework
djangorestframework-jwt
shapely
